# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi Mini Project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server Utama"
        }
    ],
    "paths": {
        "/api/v1/master/barang": {
            "get": {
                "tags": [
                    "Master Barang"
                ],
                "summary": "Get list of barang with pagination and search",
                "operationId": "1047908d14dc4dbfffc0b7e8078a98d3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by kode or nama",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Master Barang"
                ],
                "summary": "Create a new barang",
                "operationId": "d583d751c96ec875651c69a0c9cf46bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode",
                                    "nama",
                                    "kategori",
                                    "harga"
                                ],
                                "properties": {
                                    "kode": {
                                        "type": "string",
                                        "example": "B001"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Buku Tulis"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "ATK"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barang berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
...
        "/api/v1/master/barang/{id}": {
            "get": {
                "tags": [
                    "Master Barang"
                ],
                "summary": "Get a single barang by ID",
                "operationId": "349552af026fda2c43377dfff23643e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Barang not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Master Barang"
                ],
                "summary": "Update an existing barang",
                "operationId": "b2dc727f0f999e5428049465f4bf999b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode": {
                                        "type": "string",
                                        "example": "B001"
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Buku Tulis Updated"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "ATK"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate"
                    },
                    "404": {
                        "description": "Barang not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Master Barang"
                ],
                "summary": "Delete a barang",
                "operationId": "85a19b22c8733613bcb12fe5f2b57681",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of barang",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus"
                    },
                    "404": {
                        "description": "Barang not found"
                    }
                }
            }
        },
...
    },
...
}